requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pinecone-client==3.0.0
openai==1.3.0
PyPDF2==3.0.1
requests==2.31.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.1
tenacity==8.2.3
numpy==1.26.2
